backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
pymongo==4.6.0
python-dotenv==1.0.0
PyJWT==2.8.0
bcrypt==4.1.1
pandas==2.1.4
numpy>=1.26.0
joblib==1.3.2
matplotlib==3.8.2
scikit-learn==1.3.2